fix: move imageryLayerProps to type.ts
