+++ web/src/beta/lib/core/engines/Cesium/core/labels/JapanGSIOptimalBVmapVectorMapLabel/types.ts
@@ -1,7 +1,7 @@
-import { type ImageryLayer } from "@cesium/engine";
+import { ImageryLayer as CesiumImageryLayer, type ImageryProvider } from "@cesium/engine";
 
 export interface Imagery {
-  imageryLayer: ImageryLayer;
+  imageryLayer: CesiumImageryLayer;
   x: number;
   y: number;
   level: number;
@@ -11,6 +11,16 @@ export interface KeyedImagery extends Imagery {
   key: string;
   children?: KeyedImagery[];
   descendants?: KeyedImagery[];
+}
+
+export interface ImageryLayerHandle {
+  imageryLayer: CesiumImageryLayer;
+  bringToFront: () => void;
+  sendToBack: () => void;
+}
+
+export interface ImageryLayerProps extends CesiumImageryLayer.ConstructorOptions {
+  imageryProvider: ImageryProvider;
 }
 
 export type ImageryCoords = Pick<Imagery, "x" | "y" | "level">;
